more config, better code structure
