+++ minifaas-web/src/routes/mod.rs
@@ -1,0 +1,7 @@
+mod api;
+mod functions;
+mod views;
+
+pub use api::*;
+pub use views::*;
+pub use functions::*;
